feat: Adapt for TypeScript
