+++ src/views/counter.ts
@@ -1,13 +1,8 @@
 import { Listener } from "@lenra/app";
 import { Counter } from "../classes/Counter.js";
 
-/**
- * 
- * @param {Counter[]} param0 
- * @param {*} _props 
- * @returns {import("@lenra/app").JsonViewResponse}
- */
-export default function ([counter], _props) {
+export default function (counters: Counter[], _props: any) {
+  const counter = counters[0];
   return {
     value: counter.count,
     onIncrement: Listener("increment")
