--- src/views/lenra/main.js
@@ -1,13 +1,0 @@
-import { Flex, View } from "@lenra/app";
-
-export default function (_data, _props) {
-  return Flex([
-    View("lenra.menu"),
-    View("lenra.home")
-  ])
-    .direction("vertical")
-    .scroll(true)
-    .spacing(4)
-    .crossAxisAlignment("center")
-}
-
